+++ source/Examples/Resources/SqlDatabaseRole/3-EnforceDatabaseRoleMembers.ps1
@@ -25,7 +25,7 @@ Configuration Example
         {
             ServerName           = 'sqltest.company.local'
             InstanceName         = 'DSC'
-            Database             = 'AdventureWorks'
+            DatabaseName         = 'AdventureWorks'
             Name                 = 'ReportViewer'
             Members              = @('CONTOSO\Barbara', 'CONTOSO\Fred')
             Ensure               = 'Present'