+++ source/Examples/Resources/SqlDatabaseRole/6-MembersToIncludeAndExcludeInDatabaseRole.ps1
@@ -27,7 +27,7 @@ Configuration Example
         {
             ServerName           = 'sqltest.company.local'
             InstanceName         = 'DSC'
-            Database             = 'AdventureWorks'
+            DatabaseName         = 'AdventureWorks'
             Name                 = 'ReportViewer'
             MembersToInclude     = @('CONTOSO\Barbara', 'CONTOSO\Fred')
             MembersToExclude     = @('CONTOSO\Intern1')