+++ tests/Unit/MSFT_SqlDatabasePermission.Tests.ps1
@@ -116,14 +116,14 @@ try
                                                 Add-Member -MemberType NoteProperty -Name PermissionState -Value $mockSqlPermissionState -PassThru |
                                                 Add-Member -MemberType NoteProperty -Name Grantee -Value $mockExpectedSqlServerLogin -PassThru |
                                                 Add-Member -MemberType NoteProperty -Name GrantorType -Value 'User' -PassThru |
-                                                Add-Member -MemberType NoteProperty -Name ObjectClass -Value 'Database' -PassThru |
+                                                Add-Member -MemberType NoteProperty -Name ObjectClass -Value 'DatabaseName' -PassThru |
                                                 Add-Member -MemberType NoteProperty -Name ObjectName -Value $mockSqlDatabaseName -PassThru
                                             $mockEnumDatabasePermissions += New-Object -TypeName Object |
                                                 Add-Member -MemberType NoteProperty -Name PermissionType -Value $mockSqlPermissionType02 -PassThru |
                                                 Add-Member -MemberType NoteProperty -Name PermissionState -Value $mockSqlPermissionState -PassThru |
                                                 Add-Member -MemberType NoteProperty -Name Grantee -Value $mockExpectedSqlServerLogin -PassThru |
                                                 Add-Member -MemberType NoteProperty -Name GrantorType -Value 'User' -PassThru |
-                                                Add-Member -MemberType NoteProperty -Name ObjectClass -Value 'Database' -PassThru |
+                                                Add-Member -MemberType NoteProperty -Name ObjectClass -Value 'DatabaseName' -PassThru |
                                                 Add-Member -MemberType NoteProperty -Name ObjectName -Value $mockSqlDatabaseName -PassThru
 
                                             $mockEnumDatabasePermissions
@@ -246,13 +246,13 @@ try
                 It 'Should throw the correct error' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = 'unknownDatabaseName'
+                        DatabaseName    = 'unknownDatabaseName'
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
                     }
 
-                    $errorMessage = $script:localizedData.DatabaseNotFound -f $testParameters.Database
+                    $errorMessage = $script:localizedData.DatabaseNotFound -f $testParameters.DatabaseName
 
                     { Get-TargetResource @testParameters } | Should -Throw $errorMessage
 
@@ -264,7 +264,7 @@ try
                 It 'Should throw the correct error' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = 'unknownLoginName'
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
@@ -283,13 +283,13 @@ try
                     $mockInvalidOperationEnumDatabasePermissions = $true
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
                     }
 
-                    $errorMessage = $script:localizedData.FailedToEnumDatabasePermissions -f $testParameters.Name, $testParameters.Database
+                    $errorMessage = $script:localizedData.FailedToEnumDatabasePermissions -f $testParameters.Name, $testParameters.DatabaseName
 
                     { Get-TargetResource @testParameters } | Should -Throw $errorMessage
 
@@ -300,7 +300,7 @@ try
             Context 'When the system is in the desired state and ensure is set to Absent' {
                 $testParameters = $mockDefaultParameters
                 $testParameters += @{
-                    Database        = $mockSqlDatabaseName
+                    DatabaseName    = $mockSqlDatabaseName
                     Name            = $mockSqlServerLogin
                     PermissionState = 'Grant'
                     Permissions     = @( 'Connect', 'Update', 'Select' )
@@ -326,7 +326,7 @@ try
             Context 'When the system is not in the desired state and ensure is set to Absent' {
                 $testParameters = $mockDefaultParameters
                 $testParameters += @{
-                    Database        = $mockSqlDatabaseName
+                    DatabaseName    = $mockSqlDatabaseName
                     Name            = $mockSqlServerLogin
                     PermissionState = 'Grant'
                     Permissions     = @( 'Connect', 'Update' )
@@ -352,7 +352,7 @@ try
             Context 'When the system is in the desired state and ensure is set to Present' {
                 $testParameters = $mockDefaultParameters
                 $testParameters += @{
-                    Database        = $mockSqlDatabaseName
+                    DatabaseName    = $mockSqlDatabaseName
                     Name            = $mockSqlServerLogin
                     PermissionState = 'Grant'
                     Permissions     = @( 'Connect', 'Update' )
@@ -378,7 +378,7 @@ try
             Context 'When the system is not in the desired state and ensure is set to Present' {
                 $testParameters = $mockDefaultParameters
                 $testParameters += @{
-                    Database        = $mockSqlDatabaseName
+                    DatabaseName    = $mockSqlDatabaseName
                     Name            = $mockSqlServerLogin
                     PermissionState = 'Grant'
                     Permissions     = @( 'Connect', 'Update', 'Select' )
@@ -413,14 +413,14 @@ try
                 It 'Should throw the correct error' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = 'unknownDatabaseName'
+                        DatabaseName    = 'unknownDatabaseName'
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
                         Ensure          = 'Present'
                     }
 
-                    $errorMessage = $script:localizedData.DatabaseNotFound -f $testParameters.Database
+                    $errorMessage = $script:localizedData.DatabaseNotFound -f $testParameters.DatabaseName
 
                     { Test-TargetResource @testParameters } | Should -Throw $errorMessage
 
@@ -432,7 +432,7 @@ try
                 It 'Should throw the correct error' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = 'unknownLoginName'
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
@@ -452,14 +452,14 @@ try
                     $mockInvalidOperationEnumDatabasePermissions = $true
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
                         Ensure          = 'Present'
                     }
 
-                    $errorMessage = $script:localizedData.FailedToEnumDatabasePermissions -f $testParameters.Name, $testParameters.Database
+                    $errorMessage = $script:localizedData.FailedToEnumDatabasePermissions -f $testParameters.Name, $testParameters.DatabaseName
 
                     { Test-TargetResource @testParameters } | Should -Throw $errorMessage
 
@@ -471,7 +471,7 @@ try
                 It 'Should return the state as true when the desired permission does not exist' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update', 'Select' )
@@ -488,7 +488,7 @@ try
                 It 'Should return the state as false when the desired permission does not exist' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
@@ -505,7 +505,7 @@ try
                 It 'Should return the state as false when the desired permission does not exist' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update', 'Select' )
@@ -522,7 +522,7 @@ try
                 It 'Should return the state as true when the desired permission does exist' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
@@ -555,14 +555,14 @@ try
                 It 'Should throw the correct error' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = 'unknownDatabaseName'
+                        DatabaseName    = 'unknownDatabaseName'
                         Name            = $mockSqlServerLogin
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
                         Ensure          = 'Present'
                     }
 
-                    $errorMessage = $script:localizedData.DatabaseNotFound -f $testParameters.Database
+                    $errorMessage = $script:localizedData.DatabaseNotFound -f $testParameters.DatabaseName
 
                     { Set-TargetResource @testParameters } | Should -Throw $errorMessage
 
@@ -574,7 +574,7 @@ try
                 It 'Should throw the correct error' {
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = 'unknownLoginName'
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
@@ -595,14 +595,14 @@ try
                     $mockInvalidOperationForCreateMethod = $true
                     $testParameters = $mockDefaultParameters
                     $testParameters += @{
-                        Database        = $mockSqlDatabaseName
+                        DatabaseName    = $mockSqlDatabaseName
                         Name            = $mockSqlServerLoginUnknown
                         PermissionState = 'Grant'
                         Permissions     = @( 'Connect', 'Update' )
                         Ensure          = 'Present'
                     }
 
-                    $errorMessage = $script:localizedData.FailedToAddUser -f $testParameters.Name, $testParameters.Database
+                    $errorMessage = $script:localizedData.FailedToAddUser -f $testParameters.Name, $testParameters.DatabaseName
 
                     { Set-TargetResource @testParameters } | Should -Throw $errorMessage
 
@@ -623,7 +623,7 @@ try
                     It 'Should throw the correct error when mock Grant() method is called' {
                         $testParameters = $mockDefaultParameters.Clone()
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'Grant'
                             Permissions     = @( 'Connect', 'Update' )
@@ -640,7 +640,7 @@ try
                     It 'Should throw the correct error when mock Grant() method is called (for GrantWithGrant)' {
                         $testParameters = $mockDefaultParameters.Clone()
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'GrantWithGrant'
                             Permissions     = @( 'Connect', 'Update' )
@@ -658,7 +658,7 @@ try
                     It 'Should throw the correct error when mock Deny() method is called' {
                         $testParameters = $mockDefaultParameters.Clone()
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'Deny'
                             Permissions     = @( 'Connect', 'Update' )
@@ -675,7 +675,7 @@ try
                     It 'Should throw the correct error when mock Revoke() method is called' {
                         $testParameters = $mockDefaultParameters.Clone()
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'Grant'
                             Permissions     = @( 'Connect', 'Update' )
@@ -692,7 +692,7 @@ try
                     It 'Should throw the correct error when mock Revoke() method is called' {
                         $testParameters = $mockDefaultParameters.Clone()
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'GrantWithGrant'
                             Permissions     = @( 'Connect', 'Update' )
@@ -714,7 +714,7 @@ try
                             $mockExpectedSqlServerLogin = $mockSqlServerLoginUnknown
                             $testParameters = $mockDefaultParameters
                             $testParameters += @{
-                                Database        = $mockSqlDatabaseName
+                                DatabaseName    = $mockSqlDatabaseName
                                 Name            = $mockSqlServerLoginUnknown
                                 PermissionState = 'Grant'
                                 Permissions     = @( 'Connect', 'Update' )
@@ -733,7 +733,7 @@ try
                         $mockExpectedSqlServerLogin = $mockSqlServerLogin
                         $testParameters = $mockDefaultParameters
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'Grant'
                             Permissions     = @( 'Connect', 'Update' )
@@ -753,7 +753,7 @@ try
                         $mockExpectedSqlServerLogin = $mockSqlServerLogin
                         $testParameters = $mockDefaultParameters
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'GrantWithGrant'
                             Permissions     = @( 'Connect', 'Update' )
@@ -773,7 +773,7 @@ try
                         $mockExpectedSqlServerLogin = $mockSqlServerLogin
                         $testParameters = $mockDefaultParameters
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'Deny'
                             Permissions     = @( 'Connect', 'Update' )
@@ -795,7 +795,7 @@ try
                         $mockExpectedSqlServerLogin = $mockSqlServerLogin
                         $testParameters = $mockDefaultParameters
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'Grant'
                             Permissions     = @( 'Connect', 'Update' )
@@ -815,7 +815,7 @@ try
                         $mockExpectedSqlServerLogin = $mockSqlServerLogin
                         $testParameters = $mockDefaultParameters
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'GrantWithGrant'
                             Permissions     = @( 'Connect', 'Update' )
@@ -835,7 +835,7 @@ try
                         $mockExpectedSqlServerLogin = $mockSqlServerLogin
                         $testParameters = $mockDefaultParameters
                         $testParameters += @{
-                            Database        = $mockSqlDatabaseName
+                            DatabaseName    = $mockSqlDatabaseName
                             Name            = $mockSqlServerLogin
                             PermissionState = 'Deny'
                             Permissions     = @( 'Connect', 'Update' )